트렌드/mcp_online/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
PyMySQL==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.51.0
httpx==0.25.0
sqlalchemy==2.0.25
python-magic==0.4.27
